0.0.2.3.0 - Resource Modifier Optimization
Optimized Resource Modifier Applier.
Structure Upgrades.
Resource Entries actually apply now.
Don't use amend commit, this should've been 2 commits.

+++ Bribe the Birds/Assets/Scripts/Structures/StructureUpgrade.cs
@@ -15,9 +15,9 @@ public class StructureUpgrade : ScriptableObject {
     [System.Serializable]
     public struct StructureUpgradeCost {
         [SerializeField] private GameResources resource;
-        [SerializeField] private int amount;
+        [SerializeField] private float amount;
 
         public GameResources Resource { get => resource; set => resource = value; }
-        public int Amount { get => amount; set => amount = value; }
+        public float Amount { get => amount; set => amount = value; }
     }
 }
